Update FlowSensor.h datatypes
		During the tests of the library measuring running water in a residence, it was found that the variable _totalpulse periodically took negative values due an overflow.
I have made a change in the data types of the variables _pulse and _totalpulse, which were defined as "int" to type "unsigned long" to avoid overflow when reaching 32767. Now it is possible to measure up to 9.5 million liters (2^ 32-1=4E9, that is, 4294967295, which I divide by 450 pulses per liter, allows us to measure 4294967295/450=9.5444 million liters of water).